+++ std-bits/table/src/main/java/org/enso/table/write/ExcelWriter.java
@@ -15,13 +15,12 @@ import org.enso.table.data.column.storage.LongStorage;
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.table.Column;
 import org.enso.table.data.table.Table;
-import org.enso.table.error.ExistingDataException;
-import org.enso.table.error.RangeExceededException;
-import org.enso.table.error.InvalidLocationException;
+import org.enso.table.error.*;
 import org.enso.table.excel.ExcelHeaders;
 import org.enso.table.excel.ExcelRange;
 import org.enso.table.excel.ExcelRow;
 import org.enso.table.excel.ExcelSheet;
+import org.enso.table.util.ColumnMapper;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -38,12 +37,14 @@ public class ExcelWriter {
     return ensoToTextCallback;
   }
 
-  public static void getEnsoToTextCallback(Function<Object, Boolean> callback) {
-    ensoToTextCallback = callback;
+  public static void setEnsoToTextCallbackIfUnset(Function<Object, Boolean> callback) {
+    if (ensoToTextCallback == null) {
+      ensoToTextCallback = callback;
+    }
   }
 
-  public static void writeTableToSheet(Workbook workbook, int sheetIndex, boolean replace, int firstRow, Table table, Long rowLimit, ExcelHeaders.HeaderBehavior headers)
-      throws ExistingDataException, IllegalStateException {
+  public static void writeTableToSheet(Workbook workbook, int sheetIndex, ExistingDataMode existingDataMode, int firstRow, Table table, Long rowLimit, ExcelHeaders.HeaderBehavior headers)
+      throws InvalidLocationException, RangeExceededException, ExistingDataException, IllegalStateException, ColumnNameMismatchException, ColumnCountMismatchException {
     if (sheetIndex == 0 || sheetIndex > workbook.getNumberOfSheets()) {
       int i = 1;
       while (workbook.getSheet("Sheet" + i) != null) {
@@ -56,13 +57,9 @@ public class ExcelWriter {
       }
 
       writeTableToSheet(workbook, sheet, firstRow, 1, table, rowLimit, headers != ExcelHeaders.HeaderBehavior.EXCEL_COLUMN_NAMES);
-    } else if (replace) {
-      if (headers == ExcelHeaders.HeaderBehavior.INFER) {
-        ExcelSheet excelSheet = new ExcelSheet(workbook, sheetIndex);
-        headers = shouldWriteHeaders(excelSheet, firstRow + 1, 1, -1)
-            ? ExcelHeaders.HeaderBehavior.USE_FIRST_ROW_AS_HEADERS
-            : ExcelHeaders.HeaderBehavior.EXCEL_COLUMN_NAMES;
-      }
+    } else if (existingDataMode == ExistingDataMode.REPLACE) {
+      headers = headers != ExcelHeaders.HeaderBehavior.INFER ? headers :
+          shouldWriteHeaders(new ExcelSheet(workbook, sheetIndex), firstRow + 1, 1, -1);
 
       String sheetName = workbook.getSheetName(sheetIndex - 1);
       workbook.removeSheetAt(sheetIndex - 1);
@@ -70,35 +67,39 @@ public class ExcelWriter {
       Sheet sheet = workbook.createSheet(sheetName);
       workbook.setSheetOrder(sheetName, sheetIndex - 1);
       writeTableToSheet(workbook, sheet, firstRow, 1, table, rowLimit, headers != ExcelHeaders.HeaderBehavior.EXCEL_COLUMN_NAMES);
-    } else {
+    } else if (existingDataMode == ExistingDataMode.ERROR){
       throw new ExistingDataException("Sheet already exists, and cannot be replaced in current mode.");
+    } else {
+      // In Append Mode, so lets go to a Range based approach.
+      ExcelRange range = new ExcelRange(workbook.getSheetName(sheetIndex - 1), 1, 1);
+      writeTableToRange(workbook, range, existingDataMode, firstRow, table, rowLimit, headers);
     }
   }
 
-  public static void writeTableToSheet(Workbook workbook, String sheetName, boolean replace, int firstRow, Table table, Long rowLimit, ExcelHeaders.HeaderBehavior headers)
-      throws ExistingDataException, IllegalStateException {
+  public static void writeTableToSheet(Workbook workbook, String sheetName, ExistingDataMode existingDataMode, int firstRow, Table table, Long rowLimit, ExcelHeaders.HeaderBehavior headers)
+      throws InvalidLocationException, RangeExceededException, ExistingDataException, IllegalStateException, ColumnNameMismatchException, ColumnCountMismatchException {
     int sheetIndex = workbook.getNumberOfSheets() == 0 ? -1 : workbook.getSheetIndex(sheetName);
     if (sheetIndex == -1) {
       writeTableToSheet(workbook, workbook.createSheet(sheetName), firstRow, 1, table, rowLimit, headers != ExcelHeaders.HeaderBehavior.EXCEL_COLUMN_NAMES);
-    } else if (replace) {
-      if (headers == ExcelHeaders.HeaderBehavior.INFER) {
-        ExcelSheet excelSheet = new ExcelSheet(workbook, sheetIndex);
-        headers = shouldWriteHeaders(excelSheet, firstRow + 1, 1, -1)
-            ? ExcelHeaders.HeaderBehavior.USE_FIRST_ROW_AS_HEADERS
-            : ExcelHeaders.HeaderBehavior.EXCEL_COLUMN_NAMES;
-      }
+    } else if (existingDataMode == ExistingDataMode.REPLACE) {
+      headers = headers != ExcelHeaders.HeaderBehavior.INFER ? headers :
+          shouldWriteHeaders(new ExcelSheet(workbook, sheetIndex), firstRow + 1, 1, -1);
 
       workbook.removeSheetAt(sheetIndex);
       Sheet sheet = workbook.createSheet(sheetName);
       workbook.setSheetOrder(sheetName, sheetIndex);
       writeTableToSheet(workbook, sheet, firstRow, 1, table, rowLimit, headers != ExcelHeaders.HeaderBehavior.EXCEL_COLUMN_NAMES);
+    } else if (existingDataMode == ExistingDataMode.ERROR){
+      throw new ExistingDataException("Sheet already exists, and cannot be replaced in current mode.");
     } else {
-      throw new ExistingDataException("Sheet '" + sheetName + "' already exists, and cannot be replaced in current mode.");
+      // In Append Mode, so switch to appending from the top left cell of the sheet as this is equivalent to appending to the sheet.
+      ExcelRange range = new ExcelRange(sheetName, 1, 1);
+      writeTableToRange(workbook, range, existingDataMode, firstRow, table, rowLimit, headers);
     }
   }
 
-  public static void writeTableToRange(Workbook workbook, String rangeNameOrAddress, boolean replace, int skipRows, Table table, Long rowLimit, ExcelHeaders.HeaderBehavior headers)
-      throws InvalidLocationException, IllegalStateException, RangeExceededException, ExistingDataException {
+  public static void writeTableToRange(Workbook workbook, String rangeNameOrAddress, ExistingDataMode existingDataMode, int skipRows, Table table, Long rowLimit, ExcelHeaders.HeaderBehavior headers)
+      throws InvalidLocationException, IllegalStateException, RangeExceededException, ExistingDataException, ColumnNameMismatchException, ColumnCountMismatchException {
     Name name = workbook.getName(rangeNameOrAddress);
     ExcelRange excelRange;
     try {
@@ -106,11 +107,11 @@ public class ExcelWriter {
     } catch (IllegalArgumentException e) {
       throw new InvalidLocationException("Invalid range name or address '" + rangeNameOrAddress + "'.");
     }
-    writeTableToRange(workbook, excelRange, replace, skipRows, table, rowLimit, headers);
+    writeTableToRange(workbook, excelRange, existingDataMode, skipRows, table, rowLimit, headers);
   }
 
-  public static void writeTableToRange(Workbook workbook, ExcelRange range, boolean replace, int skipRows, Table table, Long rowLimit, ExcelHeaders.HeaderBehavior headers)
-      throws InvalidLocationException, IllegalStateException, RangeExceededException, ExistingDataException {
+  public static void writeTableToRange(Workbook workbook, ExcelRange range, ExistingDataMode existingDataMode, int skipRows, Table table, Long rowLimit, ExcelHeaders.HeaderBehavior headers)
+      throws InvalidLocationException, IllegalStateException, RangeExceededException, ExistingDataException, ColumnNameMismatchException, ColumnCountMismatchException {
     int sheetIndex = workbook.getSheetIndex(range.getSheetName());
     if (sheetIndex == -1) {
       throw new InvalidLocationException("Unknown sheet '" + range.getSheetName() + "'.");
@@ -127,66 +128,15 @@ public class ExcelWriter {
       }
     }
 
-    if (range.isSingleCell()) {
-      ExcelRange expanded = ExcelRange.expandSingleCell(range, sheet);
+    ExcelRange expanded = range.isSingleCell() ? ExcelRange.expandSingleCell(range, sheet) : range;
+    headers = headers != ExcelHeaders.HeaderBehavior.INFER ? headers :
+        shouldWriteHeaders(sheet, expanded.getTopRow(), expanded.getLeftColumn(), expanded.getRightColumn());
 
-      if (headers == ExcelHeaders.HeaderBehavior.INFER) {
-        headers = shouldWriteHeaders(sheet, expanded.getTopRow(), expanded.getLeftColumn(), expanded.getRightColumn())
-            ? ExcelHeaders.HeaderBehavior.USE_FIRST_ROW_AS_HEADERS
-            : ExcelHeaders.HeaderBehavior.EXCEL_COLUMN_NAMES;
-      }
-
-      // Expand to cover required size
-      int rowCount = (headers == ExcelHeaders.HeaderBehavior.USE_FIRST_ROW_AS_HEADERS ? 1 : 0) + table.rowCount();
-      if (expanded.getColumnCount() < table.getColumns().length || expanded.getRowCount() < rowCount) {
-        expanded = new ExcelRange(
-            expanded.getSheetName(),
-            expanded.getLeftColumn(),
-            expanded.getTopRow(),
-            Math.max(expanded.getRightColumn(), expanded.getLeftColumn() + table.getColumns().length - 1),
-            Math.max(expanded.getBottomRow(), expanded.getTopRow() + rowCount - 1));
-      }
-
-      checkExistingRange(workbook, expanded, replace, sheet);
+    if ((existingDataMode == ExistingDataMode.APPEND_BY_NAME || existingDataMode == ExistingDataMode.APPEND_BY_INDEX) &&
+        rangeIsNotEmpty(workbook, expanded, sheet)) {
+      appendRangeWithTable(workbook, range, existingDataMode, table, rowLimit, headers, sheet, expanded);
     } else {
-      // Check Size of Range
-      int rowCount = Math.min(Math.min(workbook.getSpreadsheetVersion().getMaxRows() - range.getTopRow() + 1, rowLimit == null ? Integer.MAX_VALUE : rowLimit.intValue()), table.rowCount());
-      if (range.getColumnCount() < table.getColumns().length || range.getRowCount() < rowCount) {
-        throw new RangeExceededException("Range is too small to fit all columns.");
-      }
-
-      if (headers == ExcelHeaders.HeaderBehavior.INFER) {
-        headers = shouldWriteHeaders(sheet, range.getTopRow(), range.getLeftColumn(), range.isWholeRow() ? -1 : range.getRightColumn())
-            ? ExcelHeaders.HeaderBehavior.USE_FIRST_ROW_AS_HEADERS
-            : ExcelHeaders.HeaderBehavior.EXCEL_COLUMN_NAMES;
-      }
-
-      checkExistingRange(workbook, range, replace, sheet);
-    }
-
-    writeTableToSheet(workbook, sheet.getSheet(), range.getTopRow() - 1, range.getLeftColumn(), table, rowLimit, headers != ExcelHeaders.HeaderBehavior.EXCEL_COLUMN_NAMES);
-  }
-
-  private static void checkExistingRange(Workbook workbook, ExcelRange range, boolean replace, ExcelSheet sheet) throws ExistingDataException {
-    int topRow = range.isWholeColumn() ? 1 : range.getTopRow();
-    int bottomRow = range.isWholeColumn() ? workbook.getSpreadsheetVersion().getMaxRows() : range.getBottomRow();
-    int leftColumn = range.isWholeRow() ? 1 : range.getLeftColumn();
-    int rightColumn = range.isWholeRow() ? workbook.getSpreadsheetVersion().getMaxColumns() : range.getRightColumn();
-
-    for (int row = topRow; row <= bottomRow; row++) {
-      ExcelRow excelRow = sheet.get(row);
-      if (excelRow != null) {
-        for (int column = leftColumn; column <= rightColumn; column++) {
-          Cell cell = excelRow.get(column);
-          if (cell != null) {
-            if (replace) {
-              cell.setBlank();
-            } else {
-              throw new ExistingDataException("Range is not empty, and cannot be replaced in current mode.");
-            }
-          }
-        }
-      }
+      updateRangeWithTable(workbook, expanded, range.isSingleCell(), existingDataMode, table, rowLimit, headers, sheet);
     }
   }
 
@@ -198,6 +148,106 @@ public class ExcelWriter {
   public static Workbook createWorkbook(boolean xls_format) {
     return xls_format ? new HSSFWorkbook() : new XSSFWorkbook();
   }
+
+  private static void appendRangeWithTable(Workbook workbook, ExcelRange range, ExistingDataMode existingDataMode, Table table, Long rowLimit, ExcelHeaders.HeaderBehavior headers, ExcelSheet sheet, ExcelRange expanded)
+      throws RangeExceededException, ExistingDataException, ColumnNameMismatchException, ColumnCountMismatchException {
+    Table mappedTable = switch (existingDataMode) {
+      case APPEND_BY_INDEX -> ColumnMapper.mapColumnsByPosition(table, expanded.getColumnCount());
+      case APPEND_BY_NAME -> {
+        if (headers == ExcelHeaders.HeaderBehavior.EXCEL_COLUMN_NAMES) {
+          throw new IllegalArgumentException("Cannot append by name when headers are not present in the existing data.");
+        }
+        String[] currentHeaders = sheet.get(expanded.getTopRow()).getCellsAsText(expanded.getLeftColumn(), expanded.getRightColumn());
+        yield ColumnMapper.mapColumnsByName(table, currentHeaders);
+      }
+      default ->
+          throw new IllegalArgumentException("Internal Error: appendRangeWithTable called with illegal existing data mode '" + existingDataMode + "'.");
+    };
+
+    if (range.isSingleCell()) {
+      int bottomRow = expanded.getBottomRow();
+      int requiredRows = Math.min(mappedTable.rowCount(), rowLimit == null ? Integer.MAX_VALUE : rowLimit.intValue());
+      expanded = new ExcelRange(expanded.getSheetName(), expanded.getLeftColumn(), bottomRow + 1, expanded.getRightColumn(), bottomRow + requiredRows);
+    } else {
+      int finalRow = expanded.getLastNonEmptyRow(sheet);
+      if (finalRow == expanded.getBottomRow()) {
+        throw new RangeExceededException("The range is already full.");
+      }
+
+      expanded = new ExcelRange(expanded.getSheetName(), expanded.getLeftColumn(), finalRow + 1, expanded.getRightColumn(), expanded.getBottomRow());
+    }
+
+    updateRangeWithTable(workbook, expanded, false, existingDataMode, mappedTable, rowLimit, ExcelHeaders.HeaderBehavior.EXCEL_COLUMN_NAMES, sheet);
+  }
+
+  private static void updateRangeWithTable(Workbook workbook, ExcelRange range, boolean singleCell, ExistingDataMode existingDataMode, Table table, Long rowLimit, ExcelHeaders.HeaderBehavior headers, ExcelSheet sheet)
+      throws RangeExceededException, ExistingDataException {
+    boolean writeHeaders = headers == ExcelHeaders.HeaderBehavior.USE_FIRST_ROW_AS_HEADERS;
+    int requiredRows = Math.min(table.rowCount(), rowLimit == null ? Integer.MAX_VALUE : rowLimit.intValue()) + (writeHeaders ? 1 : 0);
+
+    if (singleCell) {
+      range = new ExcelRange(
+          range.getSheetName(),
+          range.getLeftColumn(),
+          range.getTopRow(),
+          Math.max(range.getRightColumn(), range.getLeftColumn() + table.getColumns().length - 1),
+          Math.max(range.getBottomRow(), range.getTopRow() + requiredRows - 1));
+    }
+
+    int finalRow = range.isWholeColumn() ? workbook.getSpreadsheetVersion().getMaxRows() : range.getBottomRow();
+    int availableRows = finalRow - range.getTopRow() + 1;
+    if (range.getColumnCount() < table.getColumns().length || availableRows < requiredRows) {
+      throw new RangeExceededException("Range is too small to fit all data.");
+    }
+
+    if (existingDataMode == ExistingDataMode.REPLACE) {
+      clearRange(workbook, range, sheet);
+    } else if (rangeIsNotEmpty(workbook, range, sheet)) {
+      throw new ExistingDataException("Range is not empty, and cannot be replaced in current mode.");
+    }
+
+    writeTableToSheet(workbook, sheet.getSheet(), range.getTopRow() - 1, range.getLeftColumn(), table, rowLimit, writeHeaders);
+  }
+
+  /***
+   * Checks if a range is empty.
+   * @param workbook The workbook to check.
+   * @param range The range to check.
+   * @param sheet Sheet containing the range.
+   * @return True if range is empty and clear is False, otherwise returns False.
+   */
+  private static boolean rangeIsNotEmpty(Workbook workbook, ExcelRange range, ExcelSheet sheet) {
+    ExcelRange fullRange = range.getAbsoluteRange(workbook);
+    for (int row = fullRange.getTopRow(); row <= fullRange.getBottomRow(); row++) {
+      ExcelRow excelRow = sheet.get(row);
+      if (excelRow != null && !excelRow.isEmpty(fullRange.getLeftColumn(), fullRange.getRightColumn())) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  /***
+   * Clears a range of any content.
+   * @param workbook The workbook to clear.
+   * @param range The range to clear.
+   * @param sheet Sheet containing the range.
+   */
+  private static void clearRange(Workbook workbook, ExcelRange range, ExcelSheet sheet) {
+    ExcelRange fullRange = range.getAbsoluteRange(workbook);
+    for (int row = fullRange.getTopRow(); row <= fullRange.getBottomRow(); row++) {
+      ExcelRow excelRow = sheet.get(row);
+      if (excelRow != null) {
+        for (int column = fullRange.getLeftColumn(); column <= fullRange.getRightColumn(); column++) {
+          Cell cell = excelRow.get(column);
+          if (cell != null) {
+            cell.setBlank();
+          }
+        }
+      }
+    }
+  }
+
 
   private static void writeTableToSheet(Workbook workbook, Sheet sheet, int firstRow, int firstColumn, Table table, Long rowLimit, boolean headers)
     throws IllegalStateException {
@@ -297,28 +347,30 @@ public class ExcelWriter {
   /**
    * Determines if headers should be written for the given range in {@code INFER} mode.
    *
-   * Unlike in the {@code ExcelReader}, if empty this will default to True.
+   * Unlike in the {@code ExcelReader}, if empty this will default to having headers.
    * @param excelSheet the Excel sheet to check.
    * @param topRow top row index (1-based) of the range to check.
    * @param startCol start column index (1-based) of the range to check.
    * @param endCol end column index (1-based) of the range to check. If -1 will continue until end of row.
-   * @return true if the range has headers.
+   * @return EXCEL_COLUMN_NAMES if the range has headers, otherwise USE_FIRST_ROW_AS_HEADERS.
    */
-  private static boolean shouldWriteHeaders(ExcelSheet excelSheet, int topRow, int startCol, int endCol) {
+  private static ExcelHeaders.HeaderBehavior shouldWriteHeaders(ExcelSheet excelSheet, int topRow, int startCol, int endCol) {
     ExcelRow row = excelSheet.get(topRow);
 
-    // If the first row is missing or empty, return true as defaults to writing headers.
+    // If the first row is missing or empty, should write headers.
     if (row == null || row.isEmpty(startCol, endCol)) {
-      return true;
+      return ExcelHeaders.HeaderBehavior.USE_FIRST_ROW_AS_HEADERS;
     }
 
-    // If the first row is not empty, check if all text.
+    // If the first row is not empty but not all text, should not write headers.
     if (row.getCellsAsText(startCol, endCol) == null) {
-      return false;
+      return ExcelHeaders.HeaderBehavior.EXCEL_COLUMN_NAMES;
     }
 
-    // If the second row is missing or empty or contains text, return false.
+    // If the second row is missing, empty, or not all text, should write headers.
     ExcelRow nextRow = excelSheet.get(topRow + 1);
-    return (nextRow != null && nextRow.getCellsAsText(startCol, endCol) == null);
+    return (nextRow != null && nextRow.getCellsAsText(startCol, endCol) == null)
+        ? ExcelHeaders.HeaderBehavior.USE_FIRST_ROW_AS_HEADERS
+        : ExcelHeaders.HeaderBehavior.EXCEL_COLUMN_NAMES;
   }
 }